Expose snapshot functionality from bbr package.
...so I can get this info outside the test suite, for debugging.
